Metodo PUT para metodo implementado
