+++ src/main/java/com/andre/helpdeskback/domain/Pessoa.java
@@ -7,6 +7,8 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import org.hibernate.validator.constraints.br.CPF;
+
 import com.andre.helpdeskback.domain.enums.Perfil;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
@@ -33,6 +35,7 @@ public abstract class Pessoa implements Serializable{
     @Column(unique = true)
     protected String email;
     
+    @CPF
     @Column(unique = true)
     protected String cpf;
